[Java] Make AeronCluster track connection status.
Currently, AeronCluster does very little to track its connection status.
All applications that want to react to, for example, losing network
connectivity to the leader or the cluster shutting down, need to implement
custom logic. This change provides such logic out of the box.

If connection is lost for more than newLeaderTimeoutNs, or a message can't
be sent to a new leader for messageTimeoutNs, the client will
close itself.

For applications which call pollEgress/controlledPollEgress and check
isClosed afterward no changes are needed. For applications which use
the ingress publication or the egress subscription directly, two methods
were added: trackIngressPublicationResult and pollStateChanges.

For the new leader timeout we currently use twice the default leader heartbeat
timeout, in future we would ideally make the cluster announce the actual value.

Changed the semantics of pollEgress/controlledPollEgress return value,
hopefully not a breaking change.

The state diagram is:

stateDiagram-v2
    [*] --> CONNECTED
    CONNECTED --> AWAIT_NEW_LEADER : ingress publication NOT_CONNECTED
    CONNECTED --> AWAIT_NEW_LEADER : ingress publication CLOSED
    CONNECTED --> PENDING_CLOSE : ingress publication MAX_POSITION_EXCEEDED
    CONNECTED --> AWAIT_NEW_LEADER : egress image closed
    CONNECTED --> AWAIT_NEW_LEADER_CONNECTION : NewLeaderEvent
    CONNECTED --> PENDING_CLOSE : SessionEvent CLOSED
    CONNECTED --> CLOSED : close
    AWAIT_NEW_LEADER --> AWAIT_NEW_LEADER_CONNECTION : NewLeaderEvent
    AWAIT_NEW_LEADER --> CLOSED : timeout
    AWAIT_NEW_LEADER --> CLOSED : close
    AWAIT_NEW_LEADER_CONNECTION --> CONNECTED : ingress offer successful
    AWAIT_NEW_LEADER_CONNECTION --> CLOSED : timeout
    AWAIT_NEW_LEADER_CONNECTION --> CLOSED : close
    PENDING_CLOSE --> CLOSED
    CLOSED --> [*]