[C] Fix issue for untethered slow consumers impacting whole server (#1792)
This is related to https://github.com/aeron-io/aeron/issues/1790.

In the C aeronmd application, if the media driver is only dealing with
untethered subscriptions, it will not calculate the aggregate
subscription high water mark correctly, and so untethered subscriptions
will never get image unavailable if they are a slow consumer among fast
consumers.

This aligns aeron media driver functionality in this area with the Java
media driver.