Only send TerminationAck to the leader that requested it. (#1797)
* Only send TerminationAck to the leader that requested it.

Previously, when issuing a shutdown, via the control toggle, on the
leader of a cluster and then immediately closing its ConsensusModule, it
was possible that a new leader was elected and a follower node (in both
terms) could send a TerminationAck requested by the first leader to the
second leader. As the second leader was not expecting to receive this
TerminationAck, it would throw a NPE when attempting to access
`ConsensusModuleAgent#clusterTermination`.

Now, the follower stores the `leadershipTermId` of the leader that has
requested a termination and will only send the TerminationAck if an
election has not occurred in the interim.

I have included a test which fails ~ 50% of the time before the
changes to `ConsensusModuleAgent` are applied. It was necessary to take
fine-grained control over the progress of certain threads to recreate
the problem. Therefore, I introduced a helper class `MethodCallBlocker`,
which supports blocking a particular thread when it enters a specific
method.

* Disable test due to intermittency on mac.

This test helped to prove that a fix worked to avoid an NPE due a
shutdown followed by immediate leader close.

There was some discussion here: https://weareadaptive.slack.com/archives/C07LRBUFY83/p1746783266421059
(Sorry, it is only visible to Adaptive employees)

It may rot, as it references methods by strings for the purposes of
blocking the progress of certain threads. Therefore, we may want to
consider removing it.

* Set `terminationLeadershipTermId` whenever `terminationPosition` is set.

Currently, `terminationLeadershipTermId` is only set when the follower
receives a `terminationPosition` from the leader. It uses it to
determine whether the "terminating leader" is still the leader when
ready to terminate and deciding whether to send a termination ack.

We thought it was best that the field always be set whenever
`terminationPosition` is set, i.e., even on the leader, when handling
control signals, as it will hopefully be less surprising / easier to
rely on.